+++ app/extend/context.js
@@ -6,6 +6,7 @@ module.exports = {
   },
   // 调试函数
   debug() {
+    if (this.app.config.env !== 'local') return true; // 只在本地环境输出
     const origLimit = Error.stackTraceLimit;
     Error.stackTraceLimit = 2;
     const err = new Error();